wip: rules modularised
