+++ packages/lit-analyzer/package.json
@@ -25,13 +25,13 @@
 	},
 	"scripts": {
 		"prepare": "npm run build",
-		"build": "rollup -c",
-		"watch": "rollup -c --watch",
+		"build": "tsc",
+		"watch": "tsc -w",
 		"readme": "readme generate -i readme.blueprint.md -c readme.config.json",
 		"test": "ava"
 	},
-	"main": "lib/index.cjs.js",
-	"typings": "lib/index.cjs.d.ts",
+	"main": "lib/index.js",
+	"typings": "lib/index.d.ts",
 	"files": [
 		"/lib/",
 		"cli.js"
@@ -49,11 +49,7 @@
 	"devDependencies": {
 		"@appnest/readme": "^1.2.3",
 		"@types/node": "^10.12.24",
-		"@wessberg/rollup-plugin-ts": "^1.1.62",
 		"ava": "^2.2.0",
-		"rollup": "^1.10.1",
-		"rollup-plugin-node-resolve": "^4.0.1",
-		"rollup-plugin-replace": "^2.1.0",
 		"ts-node": "^8.3.0",
 		"tslib": "^1.9.3",
 		"typescript": "^3.5.3"
