+++ packages/lit-analyzer/src/analyze/document-analyzer/html/diagnostic/validate-html-node-attr-assignment.ts
@@ -72,12 +72,6 @@ export function validateHtmlAttrAssignment(htmlAttr: HtmlNodeAttr, request: LitA
 	const rulesResult = validateHtmlAttrAssignmentRules(htmlAttr, typeB, request);
 	if (rulesResult != null) return rulesResult;
 
-	// Validate slot attribute assignments
-	if (isRuleEnabled(config, "no-unknown-slot")) {
-		const slotResult = validateHtmlAttrSlotAssignment(htmlAttr, request);
-		if (slotResult != null) return slotResult;
-	}
-
 	// Find a corresponding target for this attribute
 	const htmlAttrTarget = htmlStore.getHtmlAttrTarget(htmlAttr);
 	if (htmlAttrTarget == null) return [];
@@ -87,12 +81,6 @@ export function validateHtmlAttrAssignment(htmlAttr: HtmlNodeAttr, request: LitA
 	//  LHS === RHS (typeA === typeB)
 	// ==========================================
 	const typeA = htmlAttrTarget.getType();
-
-	// Validate lit-html directives
-	if (isRuleEnabled(config, "no-invalid-directive-binding")) {
-		const directiveResult = validateHtmlAttrDirectiveAssignment(htmlAttr, { typeA, typeB }, request);
-		if (directiveResult != null) return directiveResult;
-	}
 
 	// Validate the types of the assignment
 	const typesResult = validateHtmlAttrAssignmentTypes(htmlAttr, { typeA, typeB }, request);
@@ -165,118 +153,6 @@ function validateHtmlAttrAssignmentRules(
 	return;
 }
 
-function validateHtmlAttrSlotAssignment(
-	htmlAttr: HtmlNodeAttr,
-	{ document, htmlStore, config }: LitAnalyzerRequest
-): LitHtmlDiagnostic[] | undefined {
-	const { assignment } = htmlAttr;
-
-	if (htmlAttr.kind !== HtmlNodeAttrKind.ATTRIBUTE || assignment == null || assignment.kind !== HtmlNodeAttrAssignmentKind.STRING) return undefined;
-
-	// Check for slots
-	if (htmlAttr.name === "slot") {
-		const parent = htmlAttr.htmlNode.parent;
-		if (parent != null) {
-			const parentHtmlTag = htmlStore.getHtmlTag(parent.tagName);
-
-			if (parentHtmlTag != null && parentHtmlTag.slots.length > 0) {
-				const slotName = assignment.value;
-				const slots = Array.from(htmlStore.getAllSlotsForTag(parentHtmlTag.tagName));
-				const matchingSlot = slots.find(slot => slot.name === slotName);
-
-				if (matchingSlot == null) {
-					const validSlotNames = slots.map(s => s.name);
-					const message =
-						validSlotNames.length === 1 && validSlotNames[0].length === 0
-							? `Invalid slot name '${slotName}'. Only the unnamed slot is valid for <${parentHtmlTag.tagName}>`
-							: `Invalid slot name '${slotName}'. Valid slot names for <${parentHtmlTag.tagName}> are: ${validSlotNames
-									.map(n => `'${n}'`)
-									.join(" | ")}`;
-
-					return [
-						{
-							kind: LitHtmlDiagnosticKind.INVALID_SLOT_NAME,
-							message,
-							validSlotNames,
-							source: "no-unknown-slot",
-							severity: litDiagnosticRuleSeverity(config, "no-unknown-slot"),
-							location: { document, ...htmlAttr.location.name }
-						}
-					];
-				}
-			}
-		}
-	}
-
-	return undefined;
-}
-
-function validateHtmlAttrAssignmentTypes(
-	htmlAttr: HtmlNodeAttr,
-	{ typeA, typeB }: { typeA: SimpleType; typeB: SimpleType },
-	request: LitAnalyzerRequest
-): LitHtmlDiagnostic[] | undefined {
-	const { assignment } = htmlAttr;
-	if (assignment == null) return undefined;
-
-	const { program, document } = request;
-
-	switch (htmlAttr.modifier) {
-		case LIT_HTML_BOOLEAN_ATTRIBUTE_MODIFIER:
-			if (isRuleEnabled(request.config, "no-invalid-boolean-binding")) {
-				// Test if the user is trying to use the ? modifier on a non-boolean type.
-				if (!isAssignableToType(typeA, { kind: SimpleTypeKind.BOOLEAN }, program)) {
-					return [
-						{
-							kind: LitHtmlDiagnosticKind.BOOL_MOD_ON_NON_BOOL,
-							message: `You are using a boolean binding on a non boolean type '${toTypeString(typeA)}'`,
-							severity: litDiagnosticRuleSeverity(request.config, "no-invalid-boolean-binding"),
-							source: "no-invalid-boolean-binding",
-							location: { document, ...htmlAttr.location.name },
-							htmlAttr,
-							typeA,
-							typeB
-						}
-					];
-				}
-			}
-			break;
-
-		case LIT_HTML_PROP_ATTRIBUTE_MODIFIER:
-			break;
-
-		case LIT_HTML_EVENT_LISTENER_ATTRIBUTE_MODIFIER:
-			break;
-
-		default: {
-			// In this case there is no modifier. Therefore:
-			const results = validateStringifiedAssignment(htmlAttr, { typeA, typeB }, request);
-			if (results != null) return results;
-		}
-	}
-
-	if (isRuleEnabled(request.config, "no-incompatible-type-binding")) {
-		// Fail if the two types are not assignable to each other.
-		// Always disable "strict" in javascript files
-		const inJsFile = request.file.fileName.endsWith(".js");
-		if (!isAssignableToType(typeA, typeB, program, inJsFile ? { strict: false } : undefined)) {
-			return [
-				{
-					kind: LitHtmlDiagnosticKind.INVALID_ATTRIBUTE_EXPRESSION_TYPE,
-					message: `Type '${toTypeString(typeB)}' is not assignable to '${toTypeString(typeA)}'`,
-					severity: litDiagnosticRuleSeverity(request.config, "no-incompatible-type-binding"),
-					source: "no-incompatible-type-binding",
-					location: { document, ...htmlAttr.location.name },
-					htmlAttr,
-					typeA,
-					typeB
-				}
-			];
-		}
-	}
-	return;
-}
-
 function validateStringifiedAssignment(
 	htmlAttr: IHtmlNodeAttr,
 	{ typeA, typeB }: { typeA: SimpleType; typeB: SimpleType },
@@ -284,42 +160,6 @@ function validateStringifiedAssignment(
 ): LitHtmlDiagnostic[] | undefined {
 	const { assignment } = htmlAttr;
 	if (assignment == null) return undefined;
-
-	if (isRuleEnabled(config, "no-complex-attribute-binding")) {
-		// Only primitive types should be allowed as "typeB" and "typeA".
-		if (!isAssignableToPrimitiveType(typeB)) {
-			return [
-				{
-					kind: LitHtmlDiagnosticKind.COMPLEX_NOT_BINDABLE_IN_ATTRIBUTE_BINDING,
-					severity: litDiagnosticRuleSeverity(config, "no-complex-attribute-binding"),
-					source: "no-complex-attribute-binding",
-					message: `You are binding a non-primitive type '${toTypeString(typeB)}'. This could result in binding the string "[object Object]".`,
-					location: { document, ...htmlAttr.location.name },
-					htmlAttr,
-					typeA,
-					typeB
-				}
-			];
-		} else if (!isAssignableToPrimitiveType(typeA)) {
-			const message = `You are assigning the primitive '${toTypeString(typeB)}' to a non-primitive type '${toTypeString(
-				typeA
-			)}'. Use '.' binding instead?`;
-
-			// Fail if the user is trying to assign a primitive value to a complex value.
-			return [
-				{
-					kind: LitHtmlDiagnosticKind.PRIMITIVE_NOT_ASSIGNABLE_TO_COMPLEX,
-					severity: litDiagnosticRuleSeverity(config, "no-complex-attribute-binding"),
-					source: "no-complex-attribute-binding",
-					message,
-					location: { document, ...htmlAttr.location.name },
-					htmlAttr,
-					typeA,
-					typeB
-				}
-			];
-		}
-	}
 
 	// Test assignments to all possible type kinds
 	const typeAIsAssignableTo = {
@@ -370,8 +210,6 @@ function validateStringifiedAssignment(
 
 		// Test if removing "undefined" from typeB would work and suggest using "ifDefined".
 		else if (isRuleEnabled(config, "no-nullable-attribute-binding") && isAssignableToSimpleTypeKind(typeB, SimpleTypeKind.UNDEFINED)) {
-			//const typeBWithoutUndefined = removeUndefinedFromType(typeB);
-			//const assignableWithoutUndefined = isAssignableToType(typeA, typeBWithoutUndefined, program);
 			return [
 				{
 					kind: LitHtmlDiagnosticKind.INVALID_ATTRIBUTE_EXPRESSION_TYPE_UNDEFINED,
@@ -436,280 +274,6 @@ function validateStringifiedAssignment(
 		}
 	}
 	return;
-}
-
-function validateHtmlAttrDirectiveAssignment(
-	htmlAttr: HtmlNodeAttr,
-	{ typeA, typeB }: { typeA: SimpleType; typeB: SimpleType },
-	request: LitAnalyzerRequest
-): LitHtmlDiagnostic[] | undefined {
-	const { assignment } = htmlAttr;
-	if (assignment == null) return undefined;
-
-	const { ts, program, document } = request;
-	const checker = program.getTypeChecker();
-
-	// Type check lit-html directives
-	if (isLitDirective(typeB)) {
-		if (assignment.kind === HtmlNodeAttrAssignmentKind.EXPRESSION && ts.isCallExpression(assignment.expression)) {
-			const functionName = assignment.expression.expression.getText();
-			const args = Array.from(assignment.expression.arguments);
-
-			switch (functionName) {
-				case "ifDefined":
-					// Example: html`<img src="${ifDefined(imageUrl)}">`;
-					// Take the argument to ifDefined and remove undefined from the type union (if possible).
-					// Then test if this result is now assignable to the attribute type.
-
-					if (args.length === 1) {
-						// "ifDefined" only has an effect on "attribute" bindings
-						if (htmlAttr.kind !== HtmlNodeAttrKind.ATTRIBUTE) {
-							return [
-								{
-									kind: LitHtmlDiagnosticKind.DIRECTIVE_NOT_ALLOWED_HERE,
-									source: "no-invalid-directive-binding",
-									severity: litDiagnosticRuleSeverity(request.config, "no-invalid-directive-binding"),
-									message: `The 'ifDefined' directive has no effect here.`,
-									location: { document, ...htmlAttr.location.name }
-								}
-							];
-						}
-
-						const returnType = toSimpleType(checker.getTypeAtLocation(args[0]), checker);
-						const returnTypeWithoutUndefined = removeUndefinedFromType(returnType);
-
-						return (
-							validateHtmlAttrAssignmentTypes(
-								htmlAttr,
-								{
-									typeA,
-									typeB: returnTypeWithoutUndefined
-								},
-								request
-							) || []
-						);
-					}
-
-					break;
-
-				case "guard":
-					// Example: html`<img src="${guard([imageUrl], () => Math.random() > 0.5 ? imageUrl : "nothing.png")}">`;
-					// Check if the return value type of the function expression given to the second parameter is assignable to the attribute.
-					if (args.length === 2) {
-						const returnFunctionType = toSimpleType(checker.getTypeAtLocation(args[1]), checker);
-
-						if (returnFunctionType.kind === SimpleTypeKind.FUNCTION) {
-							const returnType = returnFunctionType.returnType;
-
-							if (returnType == null) return [];
-
-							return (
-								validateHtmlAttrAssignmentTypes(
-									htmlAttr,
-									{
-										typeA,
-										typeB: returnType
-									},
-									request
-								) || []
-							);
-						}
-					}
-					break;
-
-				case "classMap":
-					// Report error if "classMap" is not being used on the "class" attribute.
-					if (htmlAttr.name !== "class" || htmlAttr.kind !== HtmlNodeAttrKind.ATTRIBUTE) {
-						return [
-							{
-								kind: LitHtmlDiagnosticKind.DIRECTIVE_NOT_ALLOWED_HERE,
-								message: `The 'classMap' directive can only be used in an attribute binding for the 'class' attribute`,
-								source: "no-invalid-directive-binding",
-								severity: litDiagnosticRuleSeverity(request.config, "no-invalid-directive-binding"),
-								location: { document, ...htmlAttr.location.name }
-							}
-						];
-					}
-					break;
-
-				case "styleMap":
-					// Report error if "styleMap" is not being used on the "style" attribute.
-					if (htmlAttr.name !== "style" || htmlAttr.kind !== HtmlNodeAttrKind.ATTRIBUTE) {
-						return [
-							{
-								kind: LitHtmlDiagnosticKind.DIRECTIVE_NOT_ALLOWED_HERE,
-								message: `The 'styleMap' directive can only be used in an attribute binding for the 'style' attribute`,
-								source: "no-invalid-directive-binding",
-								severity: litDiagnosticRuleSeverity(request.config, "no-invalid-directive-binding"),
-								location: { document, ...htmlAttr.location.name }
-							}
-						];
-					}
-					break;
-
-				case "unsafeHTML":
-				case "cache":
-				case "repeat":
-				case "asyncReplace":
-				case "asyncAppend":
-					// These directives can only be used within a text binding.
-					// This function validating assignments is per definition used NOT in a text binding
-					return [
-						{
-							kind: LitHtmlDiagnosticKind.DIRECTIVE_NOT_ALLOWED_HERE,
-							message: `The '${functionName}' directive can only be used within a text binding.`,
-							source: "no-invalid-directive-binding",
-							severity: litDiagnosticRuleSeverity(request.config, "no-invalid-directive-binding"),
-							location: { document, ...htmlAttr.location.name }
-						}
-					];
-			}
-		}
-
-		// Now we have an unknown (user defined) directive.
-		// Return empty array and opt out of any more type checking for this directive
-		return [];
-	}
-
-	// Make sure that "classMap" and "styleMap" directives are not used in mixed bindings.
-	else if (assignment.kind === HtmlNodeAttrAssignmentKind.MIXED) {
-		//
-		// Disabled this check because it seems that the requirement has been changed
-		//
-		// Find all relevant usage of directives in the assignment
-		/*const directivesUnavailableInMixed = assignment.values
-		 .filter((value): value is CallExpression => typeof value !== "string" && ts.isCallExpression(value))
-		 .filter(exp => ["classMap", "styleMap"].includes(exp.expression.getText()))
-		 .map(exp => ({
-		 type: toSimpleType(checker.getTypeAtLocation(exp), checker),
-		 name: exp.expression.getText()
-		 }))
-		 .filter(directive => isLitDirective(directive.type));
-
-		 if (directivesUnavailableInMixed.length > 0) {
-		 const directiveName = directivesUnavailableInMixed[0].name;
-
-		 return [
-		 {
-		 kind: LitHtmlDiagnosticKind.DIRECTIVE_NOT_ALLOWED_HERE,
-		 message: `The '${directiveName}' directive must be the entire value of the attribute.`,
-		 severity: "error",
-		 location: { document, ...htmlAttr.location.name }
-		 }
-		 ];
-		 }*/
-	}
-
-	return undefined;
-}
-
-/**
- * Checks whether a type is a lit directive.
- * It will return true if the type is a function that takes a Part type and returns a void.
- * @param type
- */
-function isLitDirective(type: SimpleType): boolean {
-	switch (type.kind) {
-		case SimpleTypeKind.ALIAS:
-			return type.name === "DirectiveFn" || isLitDirective(type.target);
-		case SimpleTypeKind.FUNCTION:
-			return (
-				type.kind === SimpleTypeKind.FUNCTION &&
-				type.argTypes != null &&
-				type.argTypes.length > 0 &&
-				["Part", "NodePart", "AttributePart", "PropertyPart"].includes(type.argTypes[0].type.name || "") &&
-				type.returnType != null &&
-				type.returnType.kind === SimpleTypeKind.VOID
-			);
-		case SimpleTypeKind.GENERIC_ARGUMENTS:
-			// Test for the built in type from lit-html: Directive<NodePart>
-			return type.target.kind === SimpleTypeKind.FUNCTION && type.target.name === "Directive";
-		default:
-			return false;
-	}
-}
-
-function removeUndefinedFromType(type: SimpleType): SimpleType {
-	switch (type.kind) {
-		case SimpleTypeKind.ALIAS:
-			return {
-				...type,
-				target: removeUndefinedFromType(type.target)
-			};
-		case SimpleTypeKind.UNION:
-			return {
-				...type,
-				types: type.types.filter(t => !isAssignableToSimpleTypeKind(t, SimpleTypeKind.UNDEFINED))
-			};
-	}
-
-	return type;
-}
-
-/**
- * Relax the type so that for example "string literal" become "string" and "function" become "any"
- * This is used for javascript files to provide type checking with Typescript type inferring
- * @param type
- */
-export function relaxType(type: SimpleType): SimpleType {
-	switch (type.kind) {
-		case SimpleTypeKind.INTERSECTION:
-		case SimpleTypeKind.UNION:
-			return {
-				...type,
-				types: type.types.map(t => relaxType(t))
-			};
-
-		case SimpleTypeKind.ENUM:
-			return {
-				...type,
-				types: type.types.map(t => relaxType(t) as SimpleTypeEnumMember)
-			};
-
-		case SimpleTypeKind.ARRAY:
-			return {
-				...type,
-				type: relaxType(type.type)
-			};
-
-		case SimpleTypeKind.PROMISE:
-			return {
-				...type,
-				type: relaxType(type.type)
-			};
-
-		case SimpleTypeKind.INTERFACE:
-		case SimpleTypeKind.OBJECT:
-		case SimpleTypeKind.FUNCTION:
-		case SimpleTypeKind.CLASS:
-			return {
-				kind: SimpleTypeKind.ANY
-			};
-
-		case SimpleTypeKind.NUMBER_LITERAL:
-			return { kind: SimpleTypeKind.NUMBER };
-		case SimpleTypeKind.STRING_LITERAL:
-			return { kind: SimpleTypeKind.STRING };
-		case SimpleTypeKind.BOOLEAN_LITERAL:
-			return { kind: SimpleTypeKind.BOOLEAN };
-		case SimpleTypeKind.BIG_INT_LITERAL:
-			return { kind: SimpleTypeKind.BIG_INT };
-
-		case SimpleTypeKind.ENUM_MEMBER:
-			return {
-				...type,
-				type: relaxType(type.type)
-			} as SimpleTypeEnumMember;
-
-		case SimpleTypeKind.ALIAS:
-			return {
-				...type,
-				target: relaxType(type.target)
-			};
-
-		default:
-			return type;
-	}
 }
 
 /**
